requirements.txt
pyyaml>=6.0
python-slugify>=8.0
pytest>=8.0
python-docx>=1.0
pdfplumber>=0.10
